// generated/src/app/api-models/claim-evaluation-rules.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimEvaluationRules {
  seqCerulId: number;
  primaryMatch: string;
  provider: string;
  procedureCode: string;
  benefitPackage: string;
  medicalDefinitions: string;
  timeframe: string;
  timeUnits: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}